requirements.txt
Flask~=2.0.0
python-dotenv